Fix warnings introduced in D8695, correct bn epsilon default value (10e-5)
Test Plan: all tests pass

Reviewers: simonl, davidn, georgep, jakeh, george, evan

Reviewed By: jakeh

Differential Revision: https://phabricator.sourcevertex.net/D8810

+++ willow/include/poponnx/op/averagepool.hpp
@@ -13,7 +13,7 @@ namespace poponnx {
 class AveragePoolOp : public HasReceptiveFieldOp {
 public:
   AveragePoolOp(const OperatorIdentifier &_opid,
-                int _countIncludePad,
+                int64_t _countIncludePad,
                 const std::vector<int64_t> &_kernelShape,
                 const HasReceptiveFieldOp::Settings &settings_);
   std::unique_ptr<Op> clone() const final;
@@ -31,7 +31,7 @@ private:
   void setSpatialK() final;
 
   std::vector<int64_t> kernelShape;
-  int countIncludePad;
+  int64_t countIncludePad;
 };
 
 class AveragePoolGradOp : public Op {
